Lexer and Parse- to debug
Denis ,look at the bug report.txt. This are the bugs I found for the
mean while in the lexer.
You better also check I didn't ruin anything with my changes (I have
tried to be careful, but there are always mistakes).

To every one- I just scratched the surface with the debug, we should
build some tests.